requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
httpx
passlib[bcrypt]
bcrypt==4.0.1
python-jose
loguru
python-multipart
PyMuPDF